Refs #16687 Extract (in)elastic parts to separate files
